4.21 is DONE

+++ controllers/blogsController.js
@@ -72,8 +72,26 @@ blogsRouter.post("/", async (request, response, next) => {
     }
 })
 
+// blogsRouter.delete("/:id", async (request, response, next) => {
+//     try {
+//         await Blog.findByIdAndRemove(request.params.id)
+//         response.status(204).end()
+//     }
+//     catch (error) {
+//         next(error)
+//     }
+// })
+
 blogsRouter.delete("/:id", async (request, response, next) => {
-    try {
+    // const body = request.body
+    //token auth
+    const token = getTokenFrom(request)
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+        return response.status(401).json({ error: "token missing or invalid" })
+    }
+
+    try{
         await Blog.findByIdAndRemove(request.params.id)
         response.status(204).end()
     }
